Don't import unicode_literals in h/util/uri.py
urllib and urlparse expect byte strings as input in Python 2 (and will
crash if given unicode), but expect unicode strings as input in Python 3.

The easiest way to handle this is to work with byte string literals in
Python 2 and to work with unicode literals in Python 3, and the easiest
way to do _that_ is to just _not_ import unicode_literals in the file.
(If we do import unicode_literals then we would have to wrap every
string literal in uri.py in a str() call just to turn them into byte
strings in Python 2.)

See:

https://github.com/hypothesis/h/pull/4807