Fix functional tests in Python 2
Adding `from __future__ import unicode_literals` to the five files that
this commit modifies caused errors from webtest about strings in HTTP
headers and values, when the functional tests are run in Python 2.

Adding str() to HTTP header and value strings fixes this in Python 2
without breaking it in Python 3.

We should revert this commit when we no longer need to support Python 2.